ref: Restructure and document the public core API
The `sentry-core` exports have been a bit restructured, other things in
`sentry` have been marked as deprecated.
