+++ sentry-error-chain/src/lib.rs
@@ -30,9 +30,9 @@ use std::fmt::{Debug, Display};
 
 use error_chain::ChainedError;
 use sentry_backtrace::backtrace_to_stacktrace;
-use sentry_core::internals::Uuid;
 use sentry_core::parse_type_from_debug;
 use sentry_core::protocol::{Event, Exception, Level};
+use sentry_core::types::Uuid;
 use sentry_core::{ClientOptions, Hub, Integration};
 
 fn exceptions_from_error_chain<T>(error: &T) -> Vec<Exception>
