+++ sentry/tests/test_client.rs
@@ -10,7 +10,7 @@ fn test_into_client() {
         let dsn = c.dsn().unwrap();
         assert_eq!(dsn.public_key(), "public");
         assert_eq!(dsn.host(), "example.com");
-        assert_eq!(dsn.scheme(), sentry::internals::Scheme::Https);
+        assert_eq!(dsn.scheme(), sentry::types::Scheme::Https);
         assert_eq!(dsn.project_id().value(), 42);
     }
 
@@ -25,7 +25,7 @@ fn test_into_client() {
         let dsn = c.dsn().unwrap();
         assert_eq!(dsn.public_key(), "public");
         assert_eq!(dsn.host(), "example.com");
-        assert_eq!(dsn.scheme(), sentry::internals::Scheme::Https);
+        assert_eq!(dsn.scheme(), sentry::types::Scheme::Https);
         assert_eq!(dsn.project_id().value(), 42);
         assert_eq!(&c.options().release.as_ref().unwrap(), &"foo@1.0");
     }
